+++ LudoGameApi/Models/GameSession.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace LudoGameApi.Objects
+namespace LudoGameApi.Models
 {
     public class GameSession
     {
