+++ LudoGameApi/Models/Player.cs
@@ -1,10 +1,11 @@
-﻿using System;
+﻿using LudoGameApi.Objects;
+using System;
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace LudoGameApi.Objects
+namespace LudoGameApi.Models
 {
     public class Player
     {
